adk-mcp/requirements.txt
google-adk
litellm
openai
fastapi
uvicorn
mcp>=1.10.1
anyio